add test for createstubs and minified
