+++ .coveragerc
@@ -26,6 +26,7 @@ exclude_lines =
     # Don't complain about missing debug-only code:
     def __repr__
     if self\.debug
+    self._log.debug
 
     # Don't complain if tests don't hit defensive assertion code:
     raise AssertionError
@@ -34,6 +35,16 @@ exclude_lines =
     # Don't complain if non-runnable code isn't run:
     if 0:
     if __name__ == .__main__.:
+
+    #
+    if __name__ == "__main__" or isMicroPython():
+    main()
+    show help()
+
+    #unreachable 
+    info["release"] = "2.0.0" 
+    if arch
+
 
 ignore_errors = True
